+++ .github/workflows/update-webmentions.yml
@@ -15,6 +15,8 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
+        with:
+          token: ${{ secrets.GH_PAT }}
 
       - name: Set up Node.js
         uses: actions/setup-node@v4
@@ -25,20 +27,36 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Build site with updated webmentions
+      - name: Fetch webmentions
         env:
           WEBMENTION_IO_TOKEN: ${{ secrets.WEBMENTION_IO_TOKEN }}
-        run: npm run build
+        run: node scripts/fetch-webmentions.js
 
       - name: Commit and push changes
+        id: commit
         run: |
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
           
+          git add src/_data/webmentions.json
           if [ -n "$(git status --porcelain)" ]; then
-            git add -A
-            git commit -m "chore: update webmentions and rebuild site"
+            git commit -m "chore: update webmentions data"
             git push
+            echo "changes_made=true" >> $GITHUB_OUTPUT
           else
             echo "No changes to commit"
+            echo "changes_made=false" >> $GITHUB_OUTPUT
           fi
+
+      - name: Trigger deploy workflow
+        if: steps.commit.outputs.changes_made == 'true'
+        uses: actions/github-script@v7
+        with:
+          github-token: ${{ secrets.GH_PAT }}
+          script: |
+            await github.rest.actions.createWorkflowDispatch({
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              workflow_id: 'deploy.yml',
+              ref: 'main'
+            })
