+++ package.json
@@ -28,9 +28,10 @@
   "devDependencies": {
     "@11ty/eleventy": "^3.1.5",
     "@11ty/eleventy-plugin-rss": "^3.0.0",
-    "@playwright/test": "^1.59.1",
-    "@tailwindcss/typography": "^0.5.19",
     "@anthropic-ai/claude-code": "^2.1.108",
+    "@playwright/test": "^1.59.1",
+    "@tailwindcss/cli": "^4.2.2",
+    "@tailwindcss/typography": "^0.5.19",
     "autoprefixer": "^10.5.0",
     "concurrently": "^9.2.1",
     "cssnano": "^7.1.5",
