+++ src/_data/webmentions.json
@@ -1,0 +1,4 @@
+{
+  "all": [],
+  "timestamp": 1776292158079
+}
